Cognito/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
python-jose==3.3.0
boto3==1.35.42
python-dotenv==1.0.1
requests==2.32.3